aws-agent/requirements.txt
mcp-agent==0.0.17
boto3
python-dotenv
PyYAML
